openuni/requirements.txt
gunicorn==20.0.4
django==3.2.1
django-cors-headers
django-heroku
Pillow
django-tinymce4-lite
cloudinary
django-import-export
razorpay
